requirements.txt
Django==1.10.1
django-extensions==1.7.4
djangorestframework==3.4.6
six==1.10.0
Werkzeug==0.11.11
psycopg2==2.6.2
